refactor: move generic unary nodes to main package
