+++ c2me-opts-dfc/src/main/java/com/ishland/c2me/opts/dfc/common/ast/integration/lithostitched/CeilBindings.java
@@ -25,10 +25,9 @@
 package com.ishland.c2me.opts.dfc.common.ast.integration.lithostitched;
 
 import com.ishland.c2me.opts.dfc.common.ast.AstEmitter;
-import com.ishland.c2me.opts.dfc.common.ast.AstNode;
 import com.ishland.c2me.opts.dfc.common.ast.FrontendRegistry;
 import com.ishland.c2me.opts.dfc.common.ast.McToAst;
-import com.ishland.c2me.opts.dfc.common.ast.integration.lithostitched.unary.CeilNode;
+import com.ishland.c2me.opts.dfc.common.ast.unary.CeilNode;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.world.gen.densityfunction.DensityFunction;
 import org.slf4j.Logger;
