+++ c2me-opts-dfc/src/main/java/com/ishland/c2me/opts/dfc/common/ast/unary/SqrtNode.java
@@ -22,20 +22,19 @@
  * THE SOFTWARE.
  */
 
-package com.ishland.c2me.opts.dfc.common.ast.integration.lithostitched.unary;
+package com.ishland.c2me.opts.dfc.common.ast.unary;
 
 import com.ishland.c2me.opts.dfc.common.ast.AstNode;
-import com.ishland.c2me.opts.dfc.common.ast.unary.AbstractUnaryNode;
 
-public class SinNode extends AbstractUnaryNode {
+public class SqrtNode extends AbstractUnaryNode {
 
-    public SinNode(AstNode operand) {
+    public SqrtNode(AstNode operand) {
         super(operand);
     }
 
     @Override
     protected AstNode newInstance(AstNode operand) {
-        return new SinNode(operand);
+        return new SqrtNode(operand);
     }
 
 }
